Antlrパーサ実装クラスの呼び出しを外部共
